requirements.txt
PyYAML==5.4.1
python-dotenv==0.14.0
-e .